refactor: add env.client to handle client env variables
