+++ src/lib/report-actions.ts
@@ -1,7 +1,7 @@
 "use server"
 
 import { createClient } from '@/utils/supabase/server'
-import { env } from "@/env";
+import { env } from "@/env.server";
 
 const API_KEY = env.AIRTABLE_API_KEY
 const BASE_ID = env.BASE_ID
