+++ src/utils/supabase/server.ts
@@ -2,7 +2,7 @@
 
 import { createServerClient } from '@supabase/ssr'
 import { cookies } from 'next/headers'
-import { env } from "@/env"
+import { env } from "@/env.server"
 
 export async function createClient() {
   const cookieStore = await cookies()
